+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/matcher/UrlMatcher.kt
@@ -91,8 +91,10 @@ class UrlMatcher {
      *
      * @param resourceURI URI of a resource to be loaded by the page
      * @param pageURI URI of the page
+     * @return a [Pair] of <Boolean, String?> the first indicates, if the URI matches and the second
+     * indicates the category of the match if available otherwise null.
      */
-    fun matches(resourceURI: String, pageURI: String): Boolean {
+    fun matches(resourceURI: String, pageURI: String): Pair<Boolean, String?> {
         return matches(Uri.parse(resourceURI), Uri.parse(pageURI))
     }
 
@@ -103,42 +105,45 @@ class UrlMatcher {
      *
      * @param resourceURI URI of a resource to be loaded by the page
      * @param pageURI URI of the page
+     * @return a [Pair] of <Boolean, String?> the first indicates, if the URI matches and the second
+     * indicates the category of the match if available otherwise null.
      */
     @Suppress("ReturnCount", "ComplexMethod")
-    fun matches(resourceURI: Uri, pageURI: Uri): Boolean {
+    fun matches(resourceURI: Uri, pageURI: Uri): Pair<Boolean, String?> {
         val resourceURLString = resourceURI.toString()
         val resourceHost = resourceURI.host
         val pageHost = pageURI.host
+        val notMatchesFound = false to null
 
         if (previouslyUnmatched.contains(resourceURLString)) {
-            return false
+            return notMatchesFound
         }
 
         if (whiteList?.contains(pageURI, resourceURI) == true) {
-            return false
+            return notMatchesFound
         }
 
         if (pageHost != null && pageHost == resourceHost) {
-            return false
+            return notMatchesFound
         }
 
         if (previouslyMatched.contains(resourceURLString)) {
-            return true
+            return true to null
         }
 
         if (resourceHost == null) {
-            return false
+            return notMatchesFound
         }
 
         for ((key, value) in categories) {
             if (enabledCategories.contains(key) && value.findNode(resourceHost.reverse()) != null) {
                 previouslyMatched.add(resourceURLString)
-                return true
+                return true to key
             }
         }
 
         previouslyUnmatched.add(resourceURLString)
-        return false
+        return notMatchesFound
     }
 
     companion object {